@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Linio\Component\Cache\Adapter;
 
@@ -6,80 +7,21 @@ use Linio\Component\Cache\Exception\KeyNotFoundException;
 
 interface AdapterInterface
 {
-    /**
-     * @param array $config
-     */
     public function __construct(array $config);
 
     /**
-     * @return bool
-     */
-    public function cacheNotFoundKeys();
-
-    /**
-     * @param string $key
-     *
      * @throws KeyNotFoundException
-     *
-     * @return mixed
      */
-    public function get($key);
+    public function get(string $key);
 
-    /**
-     * @param array $keys
-     *
-     * @return mixed[]
-     */
-    public function getMulti(array $keys);
-
-    /**
-     * @param string $key
-     * @param mixed  $value
-     *
-     * @return bool
-     */
-    public function set($key, $value);
-
-    /**
-     * @param array $keys
-     *
-     * @return bool
-     */
-    public function setMulti(array $data);
-
-    /**
-     * @param string $key
-     *
-     * @return bool
-     */
-    public function contains($key);
-
-    /**
-     * @param string $key
-     *
-     * @return bool
-     */
-    public function delete($key);
-
-    /**
-     * @param array $keys
-     *
-     * @return bool
-     */
-    public function deleteMulti(array $keys);
-
-    /**
-     * @return bool
-     */
-    public function flush();
-
-    /**
-     * @return string
-     */
-    public function getNamespace();
-
-    /**
-     * @param string $namespace
-     */
-    public function setNamespace($namespace);
+    public function getMulti(array $keys): array;
+    public function set(string $key, $value): bool;
+    public function setMulti(array $data): bool;
+    public function contains(string $key): bool;
+    public function delete(string $key): bool;
+    public function deleteMulti(array $keys): bool;
+    public function flush(): bool;
+    public function getNamespace(): string;
+    public function setNamespace(string $namespace);
+    public function cacheNotFoundKeys(): bool;
 }
