@@ -1,25 +1,20 @@
 <?php
+declare(strict_types=1);
 
 namespace Linio\Component\Cache;
 
 trait CacheAware
 {
     /**
-     * @var \Linio\Component\Cache\CacheService
+     * @var CacheService
      */
     protected $cacheService;
 
-    /**
-     * @return \Linio\Component\Cache\CacheService
-     */
-    public function getCacheService()
+    public function getCacheService(): CacheService
     {
         return $this->cacheService;
     }
 
-    /**
-     * @param \Linio\Component\Cache\CacheService $cacheService
-     */
     public function setCacheService(CacheService $cacheService)
     {
         $this->cacheService = $cacheService;
