@@ -2,8 +2,6 @@
 
 namespace Linio\Component\Cache\Exception;
 
-use Exception;
-
-class KeyNotFoundException extends Exception
+class KeyNotFoundException extends \Exception
 {
 }
